Abstracting out git repo type and using projectData module to get information about projects
