+++ lib/specifications/projectData.js
@@ -3,11 +3,9 @@
 var path = require('path');
 //var url = require('url');
 
-var Q = require('q');
-
 var fs = require('q-io/fs');
 
-var getRefInformation = require('./projectGitInteractions').getRefInformation;
+var getProject = require('./repositoryTypes/git').getProject;
 var config = require('../configuration').get();
 
 
@@ -16,8 +14,8 @@ var config = require('../configuration').get();
  *
  * @return a promise for the meta data for one project.
  */
-function getByName(projectData) {
-  return getRefInformation(projectData, projectData.currentBranchName);
+function get(projectData) {
+  return getProject(projectData);
 }
 
 
@@ -45,5 +43,5 @@ function getAll() {
 
 module.exports = {
   getAll: getAll,
-  getByName: getByName
+  get: get
 }
