+++ lib/specifications/repositoryTypes/git.js
@@ -91,7 +91,7 @@ function clone(projectData, cloneOptions) {
       return repository.getCurrentBranch();
     })
     .then(function (branch) {
-      return getRefInformation(projectData, branch.shorthand());
+      return getProjectData(projectData, branch.shorthand());
     });
 }
 
@@ -110,7 +110,7 @@ function update(projectData) {
       return repository.fetchAll(copingWithMacCertBug, true);
     })
     .then(function () {
-      return getRefInformation(projectData, projectData.currentBranchName);
+      return getProjectData(projectData, projectData.currentBranchName);
     });
 }
 
@@ -151,7 +151,7 @@ function getBranch(projectData, branchName) {
 }
 
 
-function getRefInformation(projectData, targetBranchName) {
+function getProjectData(projectData, targetBranchName) {
   // The targetBranchName is the full remote branch ref.
   // The shortBranchName is what the local branch would be called.
   var currentRef;
@@ -249,6 +249,6 @@ function getAllFiles(tree) {
 
 module.exports = {
   getProject: getProject,
-  getRefInformation: getRefInformation,
+  getProjectData: getProjectData,
   getFileContents: getFileContents
 };
