+++ routes/project.js
@@ -6,11 +6,7 @@ var router = express.Router();
 var appConfig = require('../lib/configuration').get();
 var path = require('path');
 
-var Q = require('q');
-
-var getProjectMetaDataByName = require('../lib/specifications/projectMetaData').getByName;
-var getRefInformation = require('../lib/specifications/projectGitInteractions').getRefInformation;
-var updateProject = require('../lib/specifications/getProject').update;
+var getProject = require('../lib/specifications/projectData').get;
 
 // List of available features in a project.
 router.get(/^\/([^\/]+)$/, function(req, res, next) {
@@ -19,7 +15,15 @@ router.get(/^\/([^\/]+)$/, function(req, res, next) {
   // Session variable.
   var branches = req.session.branches;
 
+  // Query param indicating that it should be attempted
+  // to check out the specified branch.
+  var targetBranchName = req.query.branch || false;
+
   var repoName = req.params[0];
+
+  if(targetBranchName) {
+    branches[repoName] = targetBranchName;
+  }
 
   // TODO: Have one place this object is created.
   var projectData = {
@@ -32,9 +36,6 @@ router.get(/^\/([^\/]+)$/, function(req, res, next) {
   // Query param causing a Git update (pull).
   var projectShouldUpdate = (req.query.update === 'true');
 
-  // Query param indicating that it should be attempted
-  // to check out the specified branch.
-  var targetBranchName = req.query.branch || false;
 
   // Render the project page and send to client.
   function render(projectData) {
@@ -55,7 +56,7 @@ router.get(/^\/([^\/]+)$/, function(req, res, next) {
 
   // If the update flag is set then branch change requests will be ingored.
   if (projectShouldUpdate) {
-    updateProject(projectData, branches[repoName])
+    getProject(projectData)
       .then(render)
       .catch(passError);
 
@@ -63,16 +64,16 @@ router.get(/^\/([^\/]+)$/, function(req, res, next) {
   // TODO: this should not know about Git refs.
   } else if (targetBranchName) {
     branches[repoName] = targetBranchName;
-    getProjectMetaDataByName(projectData)
+    getProject(projectData)
       .then(function(projectData) {
-          return getRefInformation(projectData, targetBranchName)
+          return getProject(projectData, targetBranchName)
       })
       .then(render)
       .catch(passError);
 
   // Else, generate the metadata and render the page.
   } else {
-    getProjectMetaDataByName(projectData, branches[repoName])
+    getProject(projectData, branches[repoName])
       .then(render)
       .catch(passError);
   }
